requirements.txt
requests>=2.31.0
rich>=13.0.0
dnspython>=2.4.0
python-whois>=0.9.0
PyYAML>=6.0
